ref(v22): Integration base (#2299)
Bricks for the building: this should make the `Integration` class a high
level orchestrator of the codegen.

This should organize the code for readability.
Can cause minor conflicts for #2280
