+++ src/typescript-api.ts
@@ -165,17 +165,6 @@ export const makeType = (
   return comment ? addJsDocComment(node, comment) : node;
 };
 
-/** @example type SomeOf<T> = T[keyof T]; */
-export const makeSomeOfHelper = () =>
-  makeType(
-    "SomeOf",
-    f.createIndexedAccessTypeNode(
-      f.createTypeReferenceNode("T"),
-      makeKeyOf("T"),
-    ),
-    { params: { T: undefined } },
-  );
-
 export const makePublicMethod = (
   name: ts.Identifier,
   params: ts.ParameterDeclaration[],
